refactor and fix windows \r\n
